rename folder with data
